优化路点存储数据结构
