+++ src/main/java/cn/nukkit/entity/path/NodeStore.java
@@ -1,0 +1,17 @@
+package cn.nukkit.entity.path;
+
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+public interface NodeStore {
+    void add(@NotNull Node node);
+
+    boolean has(long nodeHashCode);
+
+    @Nullable
+    Node remove(long nodeHashCode);
+
+    int size();
+
+    void clear();
+}
